refact: extract tree build a hook
